[templating] Added a template function
In a lot of cases you need to expand a Parameters via a template. I've added  a
macro:

	${template;-runbundles;@=@startlevel;,}

This will get the property `-runbundles` (in merged and decorated form) and will iterate over its entries. It will expand the template
for each entry. While expanding, it will set
@ to the key and all attributes with a '@' prefixed.

I've changed the macros slightly. Normally macros set override built in functions. However, this was not the case with macros that too parameters. Now the name of the macro
is used to do the lookup.

Also added some more test to get more coverage

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>