launcher: Refactor handling of embedded launcher
We properly move all of the embedded launcher considerations into
ProjectLauncherImpl. It should not have been in ProjectLauncher which is
an abstract base class. To do this, we had to stop ProjectLauncher
constructor from calling updateFromProject since when that method is
override by ProjectLauncherImpl, it executes on a partially constructed
ProjectLauncherImpl object.

We also change ProjectLauncherImpl to use the 2 arg constructor which
receives the launcher jar file, so we don't need to try and "find" it
with getBundles.

The pre jar is now also made part of the embedded repo
so that is normally "in-the-right-place" when we look for it using the
launcher jar in the embedded repo. If the pre jar is not next to the
launcher jar, then we are probably using a launcher from the runpath and
we extracts its pre jar to a temp file we later delete.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>