maven bnd repo: Fix timing issues when parsing archives
We need to block in IndexFile.update, while synchronized, until all
archive resources are parsed. Otherwise, repo users can see partial
state since the "snapshot" of resources is taken before the archive
resources are parsed and is not an actual snapshot since it is just a
view over a live concurrent map. This partial state intermittently
caused test failures.

So we now wait for parsing to complete before collecting the complete
resource collection into a ResourcesRepository. This is then indexed,
asynchronously, into a BridgeRepository.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>