build: Disable parallel gradle build in CI
Gradle runs each test task in a separate java process. When a test case
uses the Bnd workspace as a workspace in a test, then the new java
process will register a remotews port. So we end up with one port from
the main gradle process for the build and another for each parallel test
process which accesses the Bnd workspace as a workspace in a test. Since
the remote client uses the most recent remotews file, one test process
can use the remote workspace port from another test process. When this
happens and the other test process exits, then the test process using
its remote port will fail as the remote port is unexpectedly closed.

So we disable parallel builds in the CI until we can find a better
solution.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>