livecoding: Refactor live coding support from bnd-run-maven-plugin
We move it to a member class of ProjectLauncher: LiveCoding. This class
uses the new FileWatcher class to watch the necessary files from the
ProjectLauncher.

The LiveCoding class now tracks any files included in the bndrun and
recomputes the watch list if the bndrun or any included file changes.

We also add live coding support to the gradle plugin and the bnd run
command.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>