gradle: Refactor Bndrun task type so other task types can extend it
Also add a test case for Bndrun task type modeled on the test case for
the bnd-run-maven-plugin.

Note: It is still not supported to specify the bndrun property of a
Resolve task as the bnd.bnd file for a Bnd workspace project. Resolving
requires a Bndrun object and not a Project object.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>