gradle: Better handle --tests option when processing start parameters
When using --tests in the gradle command line, the --tests argument and
the following test name argument will appear in startParameter.taskNames.
We failed to handle this and treated the strings as possible task names.
We now recognize the --tests argument and remove it and the following
test name from the possible project names.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>