pom: Don't inherit -pom instruction properties from parent processor
The -pom instruction is parsed and added to the scoped processor for the
PomResource class. We must not inherit these properties from the parent
processor. They must come from the -pom instruction or be
passed to the PomResource constructor. Otherwise setting a `version`
property in the cnf/build.bnd or a .bnd file can confuse PomResource.

Fixes https://github.com/bndtools/bnd/issues/3381

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>