embedded-repo: Avoid multiple copies of the embedded repo
The embedded-repo.jar is now ~5.8 MB in size. We had many copies of it.
Most of the maven plugins held a copy, the gradle plugin held a copy and
we had a special fragment which held a copy for Bndtools in Eclipse.

To avoid all these copies, we modify biz.aQute.bnd.embedded-repo to be
useable in Eclipse, replacing org.bndtools.embeddedrepo, as well as to
be useable as a dependency for the maven and gradle plugins.

This change also allowed the removal of special case code in
Workspace.CachedFileRepo to handle running in junit mode.

Of course, bnd.jar will still contain its own copy of embedded-repo.jar
since it must run as a standalone jar. But all other users now share the
copy in biz.aQute.bnd.embedded-repo.jar.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>