metatype: char annotation elements must emit strings of the char
This is different than DS. The metatype spec is mute here. But the
metatype RI expects the char value in the xml attribute unlike DS. So
we defer to the RI. Bnd 4.3.0 already had made this change but I failed
to add tests for it.

Luckily, almost no one appears to use char valued elements in OCD
annotated annotations. So this behavioral change will probably not affect
anyone.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>