macro: Fix flatten properties to avoid invalid warnings
We now iterate over the processor hierarchy to find property values and
detect if they are not of String type. Previously, we might not get a
value from a Properties object and then ask the current Properties
object for the raw value. A parent Properties object may have held a
non-String value but asking the current Properties object to supply the
raw value will produce only null leading to an unhelpful warning.

We also change the computation of the key set to iterate over to use a
key set which excludes keys which are under setForceLocal. The previous
iteration could expose us to key names in parent Properties objects
which are hidden from the current Properties object.

We also make the non-String value types warnings pedantic.

Fixes https://github.com/bndtools/bnd/issues/3504

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>