junit-platform: Fix race condition in automatic()
This race condition caused random failures on JDK 11 on the CI builds.
A test bundle could have been installed between the initial request
creation and the opening of the BundleTracker. So tests would sometimes
see a test bundle fail to be processed.

We change to use the BundleTracker to create BundleSelectors for
all bundles and queue them for the main processing loop. In the main
processing loop, we process all available BundleSelectors together even
waiting briefly to see if the BundleTracker is still adding elements to
the queue.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>