[3558] Intermittent CI failures in biz.aQute.launcher.AlsoLauncherTest > testReporting_notUsingReferences #3558
It seems a number of JUnit bundles are installed that do not have a
BSN set (they are not bundles). The startlevel code assumed that all
code were bundles since they have no purpose (this JUnit code cannot
be available to anybody else as far as I can see?).

However, getBundleSymbolicName() can return null so I ignore bundles
in the startlevel code that have a null bsn.

Fixes #3558

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>