[snapshot] Difference between Bndtools and other drivers in resolve output
The output of a resolve operations is a list of -runbundles. This list
is calculated nowadays in one place (RunResolution) but unfortunately
there is a different workspace used in Bndtools and in other drivers. This
meant that Bndtools workspace resources were marked with the micro range
while these same workspaces in Gradle were marked with 'snapshot'.

A previous change added a WorkspaceRepositoryMarker to the repos that 
model the workspace projects. The new one used for Gradle added a 
WORKSPACE_NAMESPACE capability but the old one in Bndtools did not. 

This fix adds the same capability the old WorkspaceR5Repository so that
the versioning to identify the resource is now independent of the driver.

Long term solution should be to replace the WorkspaceR5Repository with
something more modern so that we have a single unified workspace 
resource repo.



Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>