[embedded activators] Framework start point
The original model had 2 phases:

* framework inited
* framework started
* IMMEDIATE=true -> before bundles were started
* IMMEDIATE=false (or not set)-> after all bundles were started

After the start level and the work done for launching Eclipse
we had the following situation:

* framework inited
* IMMEDIATE = "AFTER_FRAMEWORK_INIT"
* IMMEDIATE=true || BEFORE_BUNDLES_STARTED -> before bundles were started
* IMMEDIATE=false (or not set) || AFTER_BUNDLES_STARTED -> after all bundles were started
* framework started

This meant that there was no different between BEFORE_BUNDLES_STARTED and
AFTER_BUNDLES_STARTED. Therefore, for 4.3.1 we moved to during the meeting:

* framework inited
* IMMEDIATE = "AFTER_FRAMEWORK_INIT"
* IMMEDIATE=true || BEFORE_BUNDLES_STARTED -> before bundles were started
* framework started
* IMMEDIATE=false (or not set) || AFTER_BUNDLES_STARTED -> after all bundles were started

However, this still made it impossible to run the Embedded Activator
as before since some embedded activators assumed that the framework was
started. This was easy to fix by starting the framework from the embedded
activator. However, it is not backward compatible and embedded activators
should not touch the framework this way out of cohesion. There might be
other embedded activators.

This patch therefore makes it again compatible with 4.3 butt now with the new
option AFTER_FRAMEWORK_INIT and a proper extensible model.

* framework inited
* IMMEDIATE = "AFTER_FRAMEWORK_INIT"
* framework started
* IMMEDIATE=true || BEFORE_BUNDLES_STARTED -> before bundles were started
* IMMEDIATE=false (or not set) || AFTER_BUNDLES_STARTED -> after all bundles were started

I've attempted to describe (or define) it now better in the documentation and
put some caveats around it. I've also refactored the code a bit to have the 
decision points closer to each other.


Fixes #3582


Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>