[-export] Reconciliation of -export and export functions
Exporting of projects happened all over the place. This
is an attempt to align the -export with the existing
exporters that had ignored this instruction. It also
fixes 2 issues

## Fixes #3339 bndrun configurations cannot be exported

The executable file was made before the bundle jars
were stored on the file system. Did made it impossible to
include the just build JARs.

The processing is moved after the saving of the bundle
jars in the buildLocal method. The biz.aQute.launcher.AlsoLauncherTest.testExportWithIncludedBundlesInBndrun()
method tests for this case.

## Fixes 3178 "-runkeep:true" will not be considered during export as jar file

There was a method in Project that took a boolean keep
flag. Using this route, the keep flag was basically always
reset even if the -runkeep flag was set in the Project.
I could only find a reference in a PackageTask for ant
but I assume this is also used form Gradle?

I've deprecated this method because we should use the
existing properties, just the keep as a configuration
is pretty minimal. Worst case we could put the options
in the Run file but making a single setting a boolean
in the method prototype does not scale.

The deprecated method is adjusted. It now sets -runkeep
if either the keep flag is true or the property is set
to true. 

## More properties on the -export

Added a few properties on the -export instruction and
documented them.

Since the old way had a rater awkward output name it
was necessary to override the name with a `name` attribute.

The idea was to reconcile the exporters with this
instruction so there is a type attribute now. If neither
the name nor the type is set then the code takes on
backward compatibility mode and calculates the old name,
basically the name of the bndrun file (including the bndrun)
suffixed with .jar. However, it ran the name through
getOutputFile which did macro stuff. This is now
only used in backward compatible mode since some of
the exporters have different suffixes. So in non-backward
compatible mode the file name is not run through getOutputFile.

It is now also possible to set the bsn of the output
if the exporter supports it.

These options are tested in biz.aQute.launcher.AlsoLauncherTest.testExportOptions()

## Two export modes

The Project supported the pack() method and the
export(). The export method uses the plugin exporters and
should be the way forward. However, the pack method did a
bit more and was used by -export. To handle this in a unified
way, the ExecutableJarExporter was extended to support a new
type that matched the old -export pack() method. This became
the default type for -export.



Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>