[jareditor] Rewrite of the JAR editor
After changing the printout to the use
the bndlib JARPrinter, I found more and
more inconveniences in the JAR Editor.

This editor is basically a rewrite, it adds the following functions:

- Automatic binary/text detection
- Select text/binary
- Limited read can be switched off
- Drag files from the JAR into Eclipse or filesystem
- Open an editor in eclipse on a JAR Entry
  this works also for embedded jars
  recursively. Editors are closed if the
  parent is closed.
- Show size & modified


This rewrite is much smaller and hopefully cleaner. It uses an Eclipse File system
to map the JAR to an IFolder. Since there is no convenient place to root this folder, an arbitrary Java project is picked. A Java project because then the
decompiler works when you double click
a class file. These temporary IFolder objects are carefully deleted and are invisible.



Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>