junit-platform: Fix Sub-request to handle plain DiscoverySelector
JUnit 5.5 changed to call getSelectorsByType with the base
DiscoverySelector type which caused us to fail to substitute the OSGi
ready selectors.

We change to call the original request and then filter out the selector
we will supply. We then add the selectors we will supply to the stream.

Fixes https://github.com/bndtools/bnd/issues/3745

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>