tester: Update JUnit 4 to 4.13 and Assertj to 3.15
We provide JUnit 4 packages in biz.aQute.junit but don't want to import
the latest version of these packages in biz.aQute.tester and
biz.aQute.tester.junit-platform since we want to work with slightly
older versions of JUnit 4.

We change to use the JUnit 4 (non-OSGi) artifacts instead of servicemix
since there does not appear to be any progress in updating servicemix.

We relocate the aQute.junit packages from the biz.aQute.junit project
to the biz.aQute.tester project so can compile them against the older
JUnit 4 packages.

The latest Assertj artifact changes the Bundle-Symbolic name to
assertj-core.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>