[bndeditor] Fix for NPE while opening a file outside the workspace
When the IEditorInput is not mappable to a resource
but we can map it to a Java File, we allow inputResource
to be null.

To load the editor model, we then create a standalone
Run for the project.

This class could use some major cleanup. One wonders
if by now there is not better editors support in 
Eclipse?



Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>