Publish existing POM to Gradle Plugin Portal
This workaround makes use of the plugin-publish plugin's integration
with the java-gradle-plugin and maven-publish plugins. However, instead
of generating a POM based on the pluginMaven publication added by the
former, the task is disabled and the published POM is copied to its
destination. Thus, the publishPlugins task will pick it up and upload
it to the Gradle Plugin Portal.

Signed-off-by: Marc Philipp <mail@marcphilipp.de>