[launcher] Added a connect mode to the launcher.
The OSGi Alliance is working on a Connect specification.
This specification will allow an SPI to take over some
of the module handling in an OSGi Framework.

This change will first look if there is a Module
Connector loadable via the Service Loader. If so,
it will try to use OSGi Connect with the found
Module Connector. If anything fails, it will fall
back to the classic mode. I.e, the types of OSGi
Connect are not required to be part of the class path.

Also, a static method is added to call the launcher
externally. This method can take an optional 
Properties and/or an optional Bundle Activator. The
Bundle Activator is called before the bundles are started.

An implicit change is that the code now uses the Service Loader
for the framework as well. This code was originally 
artisan but we're now solidly on Java 8.



Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>