[launcher] Location format
The current launcher calculates a unique name for
each embedded JAR when creating an executable JAR.
This unique name is based on the name of the file
in the repo. Some of these files are just the bsn
but sometimes the full version is in there. 

Unfortunately, the launcher uses this path in the
executable JAR to create the location of the 
corresponding bundle. 

If the version is in that name, then this creates
a problem if the storage area is kept. If the version
of a bundle gets changed, it will create a new location
creating unwanted duplicates.

To generate a proper location, you need a strategy.
You could use a location that is unique for a
bsn, implying that all bundles are singletons. A sensible
strategy seems to include the major version in the 
location, allowing unique locations for incompatible
bundles.

Anyway, this patch allows the exporter to set a
strategy by specifying a format that uses macros. Available
are ${@version} and ${@bsn}.




Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>