launcher: Handle source enhanced files
Bndtools can attach source code to files for ease of debugging. These
source-enhanced-files have file names which start with '+'. When making
an executable, we should use the original files instead of the
source-enhanced-files.

I discovered this issue because a test case failed on my system because
I had had Bndtools add source to one of the files used by the test and
the extra '+' in the file name caused the test to fail.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>