[launcher] Support -executable location format
This fix was added to the 5.1 master but had
API changes. This is similar without the API
changes.

basically allows a -executable: location='${@bsn}'
instruction to set the format of the path used
in the launcher executable jar. The default 
path can contain versions and since the path
is used for the OSGi location, new bundles
are created then the path changes. So you get
the same bsn installed several times. Not good

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>