bnd: Improvements to bnd nexus sign command
We implement the --key option which was present but never implemented.
This allows the user to sign with a key other than the gpg default key.

We also add a --settings option to the nexus command to specify the
-connections-settings property to allow the user to specify connection
settings other than the default.

Documentation annotations are also added so that the bnd help command
can provide helpful output.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>