fix scope handling at dependency resolution
At dependency resolution the scope should be used to identify if an
artifact should be added or not.
It must not be used to skip the whole part of a tree as the childs of a
tree node could use a different scope then one of its predecessor.

Related to: https://github.com/bndtools/bnd/issues/3783

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>