resolve: Move resolve debug output report out of synchronized init
The debug report resulted in calling the resolver which called back
into the resolve context on another thread. Since the resolve context
was still initializing and held the init lock, the resolver could
deadlock.

So we move the debug report operation to after the
init operation completes and releases the init lock.

Fixes https://github.com/bndtools/bnd/issues/3900

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>