workspacerepo: Remove use of .index.xml files
We now always index through building Resources from the built bundles.
This avoids the need to cache and staleness of that cache.

We now defer, through the use of memoized suppliers, creating Resources
until they are actually necessary. So until you need to resolve, we can
defer this Resource creation. So this can significantly reduce Bndtools
startup time since we do not need to parse .index.xml files or bundles.
If you never use the resolver, we never need to index the workspace
bundles.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>