[closeable memoize] No null return
I think it is bad practice to return null since there is
no way to prevent it, i.e. it is outside contract. 

The caller must ensure that the Memoize is not closed 
when calling. Therefore, when it is called it is a 
violation of the contract and should be reported as an exception.

Also cleaned up the code to make it more clear what
that the Closeable Memoize can be in a closed state.

- Added an isClosed() method
- Added a 'safe' accept that runs inside the lock
  so the Memoize close is blocked until the accept 
  is done.
- test cases for this




Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>