[mavenbndrepo] Error status on multi files
The Maven Bnd Repository supports _multi_ files. A
PAR or ZIP file can be interpreted as holding other
JARs. 

When a file was detected, the container file
was added to the index with a 'loading ...' marker.

For a non-multi container, this error flag was
replaced with the normal resource information. However,
the multi container kept it there.

The new getStatus() picked this up as an error.

The fix is easy, we just add a resource for the
container.



Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>