memoize: Shorten code path for accept
If in initial state, we try to get write lock, then get value and
downgrade to read lock before calling consumer. If in initial and we
don't get write lock, then some other thread has the write lock to call
initial() and we can proceed to get the read lock.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>