resource: Add standard ordering to capabilities and requirements
addHashes had altered the encounter ordering for package capabilities
as they were copied to a HashSet and then possibly removed and replaced
in the capabilities if they were updated with hashes.

To keep a consistent ordering, we use a TreeMap to control the namespace
ordering and use "encounter" ordering on the sets. addHashes is fixes to
preserve the encounter ordering for the package capabilities.

Making an effort to have a defined ordering makes it easier when
regenerating indexes to compare with the prior revision of the index.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>