[gradle] Provide a better report for an error in generate
The previous report was just that it failed but did not contain any useful information why it failed.

I suggest we do not catch exceptions here. The Result means the code goes out of its way to not throw exceptions since it can report all anomalies.

There are also errors reported in the errors so if Gradle aborts when the project has errors, we can ignore the result error.

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>