[repository] Workspace repositories
Proposed change to the workspace repositories
confusion.

Eclipse registers a workspace repository that
listens to builds. The Bndrun files register
some repository that scans the projects. 

The problem was that we did not mandate a driver to
register a Workspace repository. This introduced
the WorkspaceRepositoryMaker interface to allow
mostly the Bndrun to check if there was a workspace
repository present. Overall a mess, especially
because before BJ cleaned this up, we were using a 
very convoluted way to get to a repository because
the bndlib had no knowledge of repositories.

I ran into this because the ClassIndex & External
Plugins use a new Workspace.getResourceRepository() method
that tries to provide a single view on the Workspace
and its repositories. Since Gradle does _not_
register as workspace repo, the code did not work
in Gradle for local projects.

However, the recent cleanup did not go all the way.
It created a much better repo for Central but it 
did not solve the problem of Bndrun. 

First, we must unify that there is _always_ a workspace
repository. Second, we should use the most efficient 
way to calculate the resource, which is at build time.

This PR does the following:

- It provides yet another Workspace repository: WorkspaceRepositoryDynamic
  which should become the _only_ one in 5.2.0
- A Memoize is created by the ResourceBuilder class 
  to calculate the resource based on the Jar.
- The SHA256 & length are now calculated during save for efficiency


Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>