[Bndtools Explorer] redirect paste action to filter text widget
All parts in the workbench have a chance to register global action
handlers.  The super class PackageExplorerPart does so during the
createPartControl method.  So if we just make sure and "reassign"
the global paste action to our own class during the BndtoolsExplorer
createPartControl, we can override the super behavior when needed.

Fixes #3801

Signed-off-by: Gregory Amerson <gregory.amerson@liferay.com>