filter: Use improved filter impl with caching of converted values
The values from the filter strings are parsed into strings. But they
often need to be converted to typed values for the actual comparison
with the input maps. This change caches the most recent conversion since
it is highly likely that the value types are highly consistent over the
various input maps. So we can save a lot of conversion of value to type.

This can be particularly beneficial for bnd.hashes which are all
numerical and we can be searching for a specific hash value across all
the indexed resources of which can have many package capabilities each
with many hash values with which to compare. It is also very useful
for version attributes which are generally always of type Version.

This is the filter impl I wrote for OSGi R8.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>