[generate] Adding main class plugin
While converting a workspace to leverage the generate
it turned out that many possible generators are already
on Maven Central as Main-Class jars.

This patch adds a secondary format, a Main-Class jar.
This should make it possible to use many generators out
of the Maven Central box.

Initially, I tried to make these command run inside
our own VM but the first one I tried did a System.exit
so these are now run external.

See the doc for the -generate instruction


Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>