[bndtools debug] Overhaul debug source path generation
This commit significantly changes the layout of the source dependencies
in the debug configuration window. In particular:
-DefaultSourceContainer is no longer used.
-BndDependencySourceContainer now contains a few intermediate nodes in
its hierarchy to group the dependencies in a way that corresponds to
the bnd(run) file - ie, separate entries for runee, runfw, runpath and
runbundles.
-runpath dependencies are now added too
-bundles displayed as bundles with the Bndtools bundle icon
-runee is handled directly by BndDependencySourceContainer and will add
the sources for the VM that is actually used for launching (as opposed
to the one that was used for compilation).

Fixes #3993

Signed-off-by: Fr Jeremy Krieg <fr.jkrieg@greekwelfaresa.org.au>