pom dependencies: Prefer maven metadata from repos
We collect any attributes provided by the repo into the container
attributes taking care not to override any attribute values already in
the container. Then when computing the -maven-dependencies instruction,
we can look for, and prefer, the maven attributes supplied by the repo.

If the jar does not come from a repo supplying maven attributes, we
fall back to processing pom.properties as before. This will be the case
for projects in the workspace since the workspace repo does not supply
maven attributes.

Fixes https://github.com/bndtools/bnd/issues/4015

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>