processor: A Processor's lastModified must consider the parent Processor
A Project's parent is the Workspace. When the Workspace is updated
(such as by editing cnf/build.bnd), the lastModified of the Workspace
is updated and so the lastModified of its Projects must also be updated.
So when the Projects rebuild, their Jar files will get the updated
lastModified time and so will appear to be changed to observers such as
Eclipse.

Fixes https://github.com/bndtools/bnd/issues/4017

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>