maven: Add support for -maven-release: sources;sourcepath="..."
Not everyone is excited about putting source files in the main jar
as is done with `-sources: true`. But they still want to release to
maven and thus need a meaningful source jar.

The `path` option to `-maven-release: sources` can help when you want
the source jar to include all the source files in one folder or jar.
But if your bundle contains classes from several projects, you will
need to collect source files from several places.

This commit adds support for the `sourcepath` option to
`-maven-release: sources`. The most common value for `sourcepath` will
be `sourcepath="${project.allsourcepath}"` which provides access to all
the source folders referenced by the project.

With the `sourcepath` option, the source jar construction process will
attempt to find a java source file for each class file in the releasing
jar by looking through the directories specified in the `sourcepath`.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>