maven: Refactor support for -maven-release: sources support
The sourcepath option is now called -sourcepath. If not specified
on the -maven-release: sources clause, then we fall back to the
project's `-sourcepath` setting. So this means, that for project's
using `-sources: false`, the source jar created at maven release will
look for source code in the project's sourcepath rather than simply
being empty.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>