macro: Improve performance of macro method lookup
In profiling the build for the org.osgi.test.cases.framework project,
I observed many thousands of macro method look ups which
failed. The Class.getMethod call throws a NoSuchMethodException to
indicate the method was not present. This occurred 5835 times in the
profiled project build. A primary example of this
is the use of the `${workspace}` macro whose `_workspace` method
is not on Project but on Workspace. So every time the macro needed to
be resolved in the project, Macro looked on the Project and
took a NoSuchMethodException and then retried on the
Workspace object. Macro method lookup failure also occurs for all
manifest headers such as `Bundle-Version`.

Macro lookup failure can happen up to 4 times as each attempt to find
the macro must look in ProjectBuilder, Project, Workspace, and then
finally Macro.

This change uses a Map to map the method names to invoker objects for
the method. We process each class for macro methods on demand and also
unreflect the Method to a MethodHandle on demand and memoize the
MethodHandle for future invocations. The cost for finding, or not
finding, a method invoker is now a map lookup and we avoid potentially
thousands of NoSuchMethodExceptions in a large build such as OSGi.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>