builder: Improve performance of system macro during a build
In profiling the build for the org.osgi.test.cases.framework project,
I observed many hundreds of system macro calls for manifest header
values such as Git-Description, and Git-SHA. This OSGi project uses
hundreds on inner jars which all look up the headers values when
calculating their manifest. I observed 825 system calls to git during
the project build. These calls took a significant toll on the build
time which was 5 minutes (measure while profiling).

Given, that during a build operation, the result of a system call with
specific arguments will not change, we now cache the results of the
system call across the graph of connected Builders (subbuilders and
child builders for make operations) to avoid repeated system calls.

This change reduced the number of system calls from 825 to 18, a 45x
reduction and brought the project build time down from around 5 min to
around 1 min (measured while profiling). This is a significant
savings for a large project build
such as the one profiled.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>