[launcher] Restart framework
Added a flag so that the framework can be restarted
in the same VM.

The -runframeworkrestart=true, sets the launch.framework.restart
launch property. If the framework is stopped via an update or
stop, and the launch.framework.restart is set, a new
FrameworkRestart Error is thrown after normal cleanup.

This exception is caught in the EmbeddedLauncher. When 
it catches this exception, it will restart the framework.

The EmbeddedLauncher keeps a counter in `launch.framework.restart.count`.

See -runframeworkrestart



Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>