[classpath] Add support for dependencies with Bundle-ClassPath
With this fix, if you add the "expand-bcp=true" attribute to a
bundle on your -buildpath, then Bndtools will extract any embedded
jars in its Bundle-ClassPath and add them to the compiler class
path. The Gradle workspace plugin also supports this syntax for build
fidelity.

If this functionality is needed in the Maven and Gradle non-workspace
plugins, they will need their own specific implementations as bnd is
not directly involved in the build dependency selection procedure in
those environments.

Fixes #2415. Fixes bndtools/bndtools#1871.

Signed-off-by: Fr Jeremy Krieg <fr.jkrieg@greekwelfaresa.org.au>