resolve test: RunResolutionTest needs latest embedded repo
We add biz.aQute.bnd.embedded-repo on the -testpath because the test
case needs the latest embedded repo for testing.

Normally we don't notice this since it uses the repo from the gradle
plugin (or Bndtools) which is generally only 1 (or a few for Bndtools)
builds old and the same version as we are testing. When this causes
issues is when we change to build a new minor version of Bnd.

Fixes https://github.com/bndtools/bnd/issues/3686

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>