memoize: Add new memoizer which memoizes once the predicate accepts
Sometimes you want to memoize something but that something may not be
ready. For example, it may return null until it can return a real value.
The new memoizer allows you to memoize a supplier along with a
predicate. The memoizer will memoize the first value from the source
supplier that is accepted by the predicate.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>