[baseline] Lost baseline error
If you export a package form another bundle and that
package has a baseline error, then this error is lost.

The bndlib project exports a package aQute.bnd.annotation.plugin
from the biz.aQute.bnd.annotation project. 

I added a Capability header to the BndPlugin and then
suddenly the bndlib was no longer built, but no errors
were marked.

After debugging it became clear that biz.aQute.bnd.annotation
was not baselined but the aQute.bnd.annotation.plugin
package caused a baseline error in bndlib.

The BaselineErrorHandler could not find 
the package when it tried to create a marker. It
therefore returned zero markers. The MarkerSupport
did not take the case into account where the
Error Handler had zero markers after it indicated
that it had a handler and returned.

This fix will create a default error on bnd when
the handler creates no markers. 

Took some effort to run this in the debugger :-(

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>