[bndtools launch] Change to honor buildBeforeLaunch preference
The old implementation didn't properly honor this preference due to the stray
! operator. Also, the call to the superclass would always return true due to
getBuildOrder() returning an empty array of projects to build instead of a null.

Calling the superclass seems redundant anyway - the BndtoolsBuilder incremental
build on the workspace should be smart enough to build what needs to be built.

Fixes #4089.

Signed-off-by: Fr Jeremy Krieg <fr.jkrieg@greekwelfaresa.org.au>