central: Rework callbacks for workspace creation
We now have two callback queues. One for the first workspace which may
not be a Bnd Workspace but may be a default workspace. And another for
each Bnd Workspace.

This allows m2e code to use the first workspace callback queue
(onAnyWorkspace) since the underlying promise will be resolved on the
first call to getWorkspace.

Code for Bnd Workspace models builds can use the second workspace
callback queue (onCnfWorkspace) whose underlying promise is not
resolved until a Bnd Workspace is created. If the workspace converts
back to a default workspace (when the cnf project is deleted), then a
new Bnd Workspace queue is created waiting for the workspace to get
a cnf project again.

This is not a complete replacement of the getWorkspace code in Central
but seems to be the least disruptive change to accomplish the main
goals given that we have to support a hybrid world where there must
be a workspace that may or may not be a Bnd Workspace.

Fixes https://github.com/bndtools/bnd/issues/4108

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>